pyproject.toml
[project]
name = "procesally-backend"
version = "0.1.0"
description = "Backend MVP para Procesally con FastAPI y SQLModel"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.111.0",
    "pgvector>=0.3.1",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.1",
    "sentence-transformers>=3.0.1",
    "sqlmodel>=0.0.21",
    "uvicorn[standard]>=0.30.1",
    "pydantic-settings>=2.6.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "greenlet>=3.3.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.1"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
testpaths = ["tests"]
python_files = ["test_*.py"]
pythonpath = ["."]
